Added support for missed requirment to take in a number that has a decimal in it.

+++ Formatster.Core/BillionNumberFormatter.cs
@@ -2,7 +2,7 @@ namespace Formatster.Core
 {
     public class BillionNumberFormatter : INumberFormatter
     {
-        public string Format(long numberToFormat)
+        public string Format(double numberToFormat)
         {
             return (numberToFormat / 1000000000D).ToString("0.#") + "B";
         }
